+++ src/main/java/com/LukeTheDuke9801/progressivetechnicalities/init/BlockInit.java
@@ -1,22 +1,24 @@
 package com.LukeTheDuke9801.progressivetechnicalities.init;
 
 import com.LukeTheDuke9801.progressivetechnicalities.ProgressiveTechnicalities;
-import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.ArmorDisassembler;
 import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.BlockMultiTNT;
 import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.BlockQuarry;
-import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.CastingTableBlock;
-import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.DisenchanterBlock;
 import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.FeyPortalBlock;
-import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.ItemRepairerBlock;
 import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.LargeChestBlock;
 import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.OilPortalBlock;
-import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.machines.alloy_table.AlloyTableBlock;
-import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.machines.auto_spawner.AutoSpawnerBlock;
-import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.machines.electric_furnace.ElectricFurnaceBlock;
+import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.generators.coal.CoalGeneratorBlock;
+import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.generators.lava.LavaGeneratorBlock;
+import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.generators.oil.OilGeneratorBlock;
 import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.machines.enchantment_forge.EnchantmentForgeBlock;
 import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.machines.mob_slayer.MobSlayerBlock;
-import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.machines.oil_generator.OilGeneratorBlock;
+import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.machines.mob_spawner.MobSpawnerBlock;
 import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.machines.tinker_table.TinkerTableBlock;
+import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.simple_machines.ArmorDisassembler;
+import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.simple_machines.CastingTableBlock;
+import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.simple_machines.DisenchanterBlock;
+import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.simple_machines.InductionSmelterBlock;
+import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.simple_machines.ItemRepairerBlock;
+import com.LukeTheDuke9801.progressivetechnicalities.objects.blocks.simple_machines.PulverizerBlock;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.SoundType;
@@ -42,7 +44,7 @@ public class BlockInit {
 			() -> new Block(Block.Properties.create(Material.IRON).hardnessAndResistance(4f, 6f).sound(SoundType.METAL).harvestLevel(4).harvestTool(ToolType.PICKAXE)));
 	
 	public static final RegistryObject<Block> CHROMIUM_ORE = BLOCKS.register("chromium_ore",
-			() -> new Block(Block.Properties.create(Material.IRON).hardnessAndResistance(3f, 6f).sound(SoundType.METAL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
+			() -> new Block(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 6f).sound(SoundType.METAL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
 	
 	public static final RegistryObject<Block> SKY_GEM_ORE = BLOCKS.register("sky_gem_ore",
 			() -> new Block(Block.Properties.create(Material.IRON).hardnessAndResistance(3f, 6f).sound(SoundType.METAL).harvestLevel(4).harvestTool(ToolType.PICKAXE)));
@@ -90,14 +92,14 @@ public class BlockInit {
 	
 	// Machines
 	
-	public static final RegistryObject<Block> ELECTRIC_FURNACE = BLOCKS.register("electric_furnace",
-			() -> new ElectricFurnaceBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
-	
 	public static final RegistryObject<Block> OIL_GENERATOR = BLOCKS.register("oil_generator",
 			() -> new OilGeneratorBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
 	
-	public static final RegistryObject<Block> ALLOY_TABLE = BLOCKS.register("alloy_table",
-			() -> new AlloyTableBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
+	public static final RegistryObject<Block> LAVA_GENERATOR = BLOCKS.register("lava_generator",
+			() -> new LavaGeneratorBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
+	
+	public static final RegistryObject<Block> COAL_GENERATOR = BLOCKS.register("coal_generator",
+			() -> new CoalGeneratorBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
 	
 	public static final RegistryObject<Block> CASTING_TABLE = BLOCKS.register("casting_table",
 			() -> new CastingTableBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
@@ -120,7 +122,13 @@ public class BlockInit {
 	public static final RegistryObject<Block> MOB_SLAYER = BLOCKS.register("mob_slayer",
 			() -> new MobSlayerBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
 	
-	public static final RegistryObject<Block> AUTO_SPAWNER = BLOCKS.register("auto_spawner",
-			() -> new AutoSpawnerBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
+	public static final RegistryObject<Block> MOB_SPAWNER = BLOCKS.register("mob_spawner",
+			() -> new MobSpawnerBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
+	
+	public static final RegistryObject<Block> PULVERIZER = BLOCKS.register("pulverizer",
+			() -> new PulverizerBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
+	
+	public static final RegistryObject<Block> INDUCTION_SMELTER = BLOCKS.register("induction_smelter",
+			() -> new InductionSmelterBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(2f, 1000f).sound(SoundType.ANVIL).harvestLevel(0).harvestTool(ToolType.PICKAXE)));
 	
 }