+++ src/main/java/com/LukeTheDuke9801/progressivetechnicalities/objects/items/tools/UnobtaniumSword.java
@@ -11,6 +11,7 @@ import net.minecraft.item.IItemTier;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.SwordItem;
+import net.minecraft.util.DamageSource;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.StringTextComponent;
 import net.minecraft.world.Explosion;
@@ -24,7 +25,7 @@ public class UnobtaniumSword extends SwordItem {
 	@Override
 	public void addInformation(ItemStack stack, World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
 		if (KeyboardHelper.isHoldingShift()) {
-			tooltip.add(new StringTextComponent("Kills anything in one hit but also releases a destructive shockwave"));
+			tooltip.add(new StringTextComponent("Deals 1/4 of the target's max health and bypasses armor"));
 		}
 		
 		super.addInformation(stack, worldIn, tooltip, flagIn);
@@ -32,10 +33,8 @@ public class UnobtaniumSword extends SwordItem {
 	
 	@Override
 	public boolean hitEntity(ItemStack stack, LivingEntity target, LivingEntity attacker) {
-	      target.onKillCommand();
-	      
-	      attacker.getEntityWorld().createExplosion((Entity)null, target.getPosX(), target.getPosY(), target.getPosZ(), 20f, false, Explosion.Mode.DESTROY);
-			
+	      int damage = (int) (target.getMaxHealth() / 4);
+	      target.attackEntityFrom(DamageSource.OUT_OF_WORLD, damage);
 	      return true;
 	}